feat: move ligand_prep into proteome_scan subpackage
Relocates the ligand preparation primitive from the flat
deepchem_server/core/primitives/ layout into the proteome_scan
subpackage and updates the API path accordingly.

Changes:
- Remove deepchem_server/core/primitives/ligand_prep.py
- Add deepchem_server/core/primitives/proteome_scan/ligand_prep.py
- Update /primitive/ligand-prep → /primitive/proteome-scan/ligand-prep
- Update compute.py import path
- Update test import path
